Fix build warnings and change from int to float where more precision makes sense.

+++ src/ECS/PolygonTransformComponent.hpp
@@ -10,7 +10,7 @@ struct PolygonTransformComponent : public Component
 	Vector2D velocity;
 
     SDL_FPoint* points;
-    int pointCount;
+    size_t pointCount;
 	int scale;
 	int speed;
